:truck: chore(api) move batch usenamer password generate to src
