+++ api/src/prescription/organization-learner/domain/usecases/index.js
@@ -1,12 +1,11 @@
 import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 
-import * as campaignRepository from '../../../../../src/prescription/campaign/infrastructure/repositories/campaign-repository.js';
-import * as userReconciliationService from '../../../../../src/shared/domain/services/user-reconciliation-service.js';
 import * as passwordGenerator from '../../../../identity-access-management/domain/services/password-generator.service.js';
 import * as authenticationMethodRepository from '../../../../identity-access-management/infrastructure/repositories/authentication-method.repository.js';
 import { emailValidationDemandRepository } from '../../../../identity-access-management/infrastructure/repositories/email-validation-demand.repository.js';
 import { lastUserApplicationConnectionsRepository } from '../../../../identity-access-management/infrastructure/repositories/last-user-application-connections.repository.js';
+import { organizationLearnerIdentityRepository } from '../../../../identity-access-management/infrastructure/repositories/organization-learner-identity.repository.js';
 import * as userRepository from '../../../../identity-access-management/infrastructure/repositories/user.repository.js';
 import { userToCreateRepository } from '../../../../identity-access-management/infrastructure/repositories/user-to-create.repository.js';
 import * as organizationFeaturesAPI from '../../../../organizational-entities/application/api/organization-features-api.js';
@@ -14,6 +13,7 @@ import { tagRepository } from '../../../../organizational-entities/infrastructur
 import { cryptoService } from '../../../../shared/domain/services/crypto-service.js';
 import * as obfuscationService from '../../../../shared/domain/services/obfuscation-service.js';
 import { tokenService } from '../../../../shared/domain/services/token-service.js';
+import * as userReconciliationService from '../../../../shared/domain/services/user-reconciliation-service.js';
 import * as userService from '../../../../shared/domain/services/user-service.js';
 /** TODO
  * Internal API Needed For
@@ -25,9 +25,11 @@ import * as userValidator from '../../../../shared/domain/validators/user-valida
 import * as libOrganizationLearnerRepository from '../../../../shared/infrastructure/repositories/organization-learner-repository.js';
 import * as organizationRepository from '../../../../shared/infrastructure/repositories/organization-repository.js';
 import * as userLoginRepository from '../../../../shared/infrastructure/repositories/user-login-repository.js';
+import * as writeCsvUtils from '../../../../shared/infrastructure/utils/csv/write-csv-utils.js';
 import { injectDependencies } from '../../../../shared/infrastructure/utils/dependency-injection.js';
 import { importNamedExportsFromDirectory } from '../../../../shared/infrastructure/utils/import-named-exports-from-directory.js';
 import * as emailRepository from '../../../../shared/mail/infrastructure/repositories/email.repository.js';
+import * as campaignRepository from '../../../campaign/infrastructure/repositories/campaign-repository.js';
 import * as divisionRepository from '../../../campaign/infrastructure/repositories/division-repository.js';
 import * as groupRepository from '../../../campaign/infrastructure/repositories/group-repository.js';
 import * as campaignParticipationOverviewRepository from '../../../campaign-participation/infrastructure/repositories/campaign-participation-overview-repository.js';
@@ -58,6 +60,7 @@ const dependencies = {
   organizationLearnerActivityRepository,
   organizationLearnerRepository: repositories.organizationLearnerRepository,
   organizationLearnerFeatureRepository,
+  organizationLearnerIdentityRepository,
   organizationLearnerImportFormatRepository,
   organizationFeaturesAPI,
   passwordGenerator,
@@ -77,6 +80,7 @@ const dependencies = {
   obfuscationService,
   tokenService,
   passwordValidator,
+  writeCsvUtils,
 };
 
 const path = dirname(fileURLToPath(import.meta.url));
