+++ api/tests/prescription/organization-learner/unit/domain/usecases/generate-reset-organization-learners-password-csv_test.js
@@ -1,6 +1,6 @@
-import { generateResetOrganizationLearnersPasswordCsvContent } from '../../../../lib/domain/usecases/generate-reset-organization-learners-password-cvs-content.js';
-import { OrganizationLearnerPasswordResetDTO } from '../../../../src/shared/domain/models/OrganizationLearnerPasswordResetDTO.js';
-import { expect, sinon } from '../../../test-helper.js';
+import { generateResetOrganizationLearnersPasswordCsvContent } from '../../../../../../src/prescription/organization-learner/domain/usecases/generate-reset-organization-learners-password-cvs-content.js';
+import { OrganizationLearnerPasswordResetDTO } from '../../../../../../src/shared/domain/models/OrganizationLearnerPasswordResetDTO.js';
+import { expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | UseCases | Generate reset organization learners csv', function () {
   it('returns generated CSV content', async function () {
